+++ src/Contracts/CouponInterface.php
@@ -17,10 +17,17 @@ interface CouponInterface
     public function isRedeemable(TotalingInterface $totalingItem): bool;
 
     /**
+     * How much this coupon takes off, in cents.
+     *
+     * A coupon expressed as a percentage rounds with
+     * {@see \Tnt\Ecommerce\Money::percentageOf()}, once, on the amount the
+     * percentage applies to.
+     *
+     * @see \Tnt\Ecommerce\Money
      * @param TotalingInterface $totalingItem
-     * @return float
+     * @return int
      */
-    public function getReduction(TotalingInterface $totalingItem): float;
+    public function getReduction(TotalingInterface $totalingItem): int;
 
     /**
      * @param Order $order
